# Conflicts:
#	src/Services/RoleService.php

+++ src/Services/RoleService.php
@@ -7,6 +7,7 @@ use Skoro\AdminPack\Events\RoleDeletedEvent;
 use Skoro\AdminPack\Events\RoleUpdatedEvent;
 use Skoro\AdminPack\Models\Role;
 use Skoro\AdminPack\Models\User;
+use RuntimeException;
 
 /**
  * Role create/update/delete service.
@@ -51,10 +52,17 @@ class RoleService
      * Deletes the role.
      *
      * @param Role $role       The role to delete.
-     * @param User $deletedBy  The user that is deleted the role.
+     * @param User $deletedBy  The user which is deleted the role.
+     *
+     * @throws RuntimeException When role cannot be deleted because of users.
      */
     public function delete(Role $role, User $deletedBy): bool
     {
+        $users = $role->users()->count();
+        if ($users > 0) {
+            throw new RuntimeException('Role is used by users and cannot be deleted.');
+        }
+
         $status = $role->delete();
         if ($status) {
             event(new RoleDeletedEvent($deletedBy, $role));
@@ -69,11 +77,18 @@ class RoleService
      * @param Role   $role        The role model to save.
      * @param string $name        The new role name.
      * @param int[]  $permissions The list of permission IDs.
+     *
+     * @throws RuntimeException   When the list of permissions is empty.
      */
     protected function saveRole(Role $role, string $name, array $permissions): void
     {
+        if (empty($permissions)) {
+            throw new RuntimeException('Permissions list cannot be empty.');
+        }
+
         $role->name = $name;
         $role->saveOrFail();
+
         $role->permissions()->sync($permissions);
     }
 }
